ADDED experiment settings to dssolver

+++ src/graph/graph.cpp
@@ -4,10 +4,10 @@
 #include <random>
 #include <algorithm>
 
-std::map<int, int> Graph::init(int n_, int m_, std::vector<std::pair<int, int> > edges_) {
+std::map<size_t, size_t> Graph::init(size_t n_, size_t m_, std::vector<std::pair<size_t, size_t> > edges_) {
 
     n = n_, m = m_;
-    edges = std::vector<std::pair<int, int> >(edges_.size());
+    edges = std::vector<std::pair<size_t, size_t> >(edges_.size());
 
     // std::vector<int> ver;
     // for (int e = 0; e < m; ++e) {
@@ -17,29 +17,29 @@ std::map<int, int> Graph::init(int n_, int m_, std::vector<std::pair<int, int> >
     // }
     // sort(ver.begin(), ver.end());
     // ver.erase(unique(ver.begin(), ver.end()), ver.end());
-    std::map<int, int> ids;
+    std::map<size_t, size_t> ids;
     // for (int i = 0; i < n; ++i) {
     //     ids[ver[i]] = i;
     // }
-    for (int e = 0; e < m; ++e) {
+    for (size_t e = 0; e < m; ++e) {
         auto [u, v] = edges_[e];
         if (ids.find(u) == ids.end()) {
-            int tot = (int)ids.size();
+            size_t tot = ids.size();
             ids[u] = tot;
         }
         if (ids.find(v) == ids.end()) {
-            int tot = (int)ids.size();
+            size_t tot = ids.size();
             ids[v] = tot;
         }
     }
 
-    n_ = (int)ids.size();
+    n_ = ids.size();
 
     std::mt19937 rnd(2226701);
     shuffle(edges.begin(), edges.end(), rnd);
     
-    g = std::vector<std::vector<std::pair<int, int> > > (n, std::vector<std::pair<int, int> >());
-    for (int e = 0; e < m; ++e) {
+    g = std::vector<std::vector<std::pair<size_t, size_t> > > (n, std::vector<std::pair<size_t, size_t> >());
+    for (size_t e = 0; e < m; ++e) {
         auto [u, v] = edges_[e];
         u = ids[u], v = ids[v];
         edges[e] = std::make_pair(u, v);
@@ -54,11 +54,11 @@ std::map<int, int> Graph::init(int n_, int m_, std::vector<std::pair<int, int> >
 void Graph::readFromText(const char *input_path) {
     FILE *input_file = fopen(input_path, "r");
 
-    int n, m;
-    fscanf(input_file, "%d%d", &n, &m);
-    std::vector<std::pair<int, int> > edges(m);
-    for (int e = 0; e < m; ++e) {
-        int u, v; fscanf(input_file, "%d%d", &u, &v);
+    size_t n, m;
+    fscanf(input_file, "%zu%zu", &n, &m);
+    std::vector<std::pair<size_t, size_t> > edges(m);
+    for (size_t e = 0; e < m; ++e) {
+        size_t u, v; fscanf(input_file, "%zu%zu", &u, &v);
         edges[e] = std::make_pair(u, v);
     }
     fclose(input_file);
@@ -69,65 +69,65 @@ void Graph::readFromText(const char *input_path) {
 void Graph::writeToText(const char *output_path) const {
     FILE *output_file = fopen(output_path, "w");
 
-    fprintf(output_file, "%d %d\n", n, m);
+    fprintf(output_file, "%zu %zu\n", n, m);
     for (auto [u, v]: edges) {
-        fprintf(output_file, "%d %d\n", u, v);
+        fprintf(output_file, "%zu %zu\n", u, v);
     }
     fclose(output_file);
 }
 
-Graph *Graph::induced_subgraph(const VertexSet &V, const std::vector<int> &mapping, VertexSet *R, VertexSet *A) const {
+Graph *Graph::induced_subgraph(const VertexSet &V, const std::vector<size_t> &mapping, VertexSet *R, VertexSet *A) const {
     Graph *G_ = new Graph();
 
-    std::vector<std::pair<int, int> > edges_;
+    std::vector<std::pair<size_t, size_t> > edges_;
     for (auto e: edges) {
         if (V.in[e.first] && V.in[e.second]) {
             edges_.push_back(std::make_pair(mapping[e.first], mapping[e.second]));
         }
     }
 
-    std::map<int, int> ord = (G_ -> init((int)V.list.size(), (int)edges_.size(), edges_));
+    std::map<size_t, size_t> ord = (G_ -> init(V.list.size(), edges_.size(), edges_));
     R -> reorder(ord);
     A -> reorder(ord);
     return G_;
 }
 
 void Graph::shrink(VertexSet *R, VertexSet *A) {
-    std::vector<std::pair<int, int> > edges_;
+    std::vector<std::pair<size_t, size_t> > edges_;
     for (auto e: edges) {
         if ((R -> in)[e.first] || (R -> in)[e.second]) {
             edges_.push_back(e);
         }
     }
-    std::map<int, int> ord = init(n, (int)(edges_.size()), edges);
+    std::map<size_t, size_t> ord = init(n, edges_.size(), edges);
     R -> reorder(ord);
     A -> reorder(ord);
 }
 
-void VertexSet::init(int n_, std::vector<int> list_) {
+void VertexSet::init(size_t n_, std::vector<size_t> list_) {
     n = n_;
     list = list_;
-    in = std::vector<int>(n, 0);
+    in = std::vector<size_t>(n, 0);
     for (auto u: list) in[u] = true;
-    eprintf("VERTEXSET size = %d\n", (int)size());
+    eprintf("VERTEXSET size = %zu\n", size());
     // for (auto u: list) eprintf("%d ", u); eputs("");
 }
 
 size_t VertexSet::size() const{ return list.size(); }
 
-VertexSet *VertexSet::induced_list(const VertexSet &V, const std::vector<int> &mapping) const {
+VertexSet *VertexSet::induced_list(const VertexSet &V, const std::vector<size_t> &mapping) const {
     VertexSet *V_ = new VertexSet();
-    std::vector<int> list_;
+    std::vector<size_t> list_;
     for (auto u: list) {
         if (V.in[u]) {
             list_.push_back(mapping[u]);
         }
     }
-    V_ -> init((int)(V.size()), list_);
+    V_ -> init(V.size(), list_);
     return V_;
 }
 
-void VertexSet::reorder(std::map<int, int> mapping) {
+void VertexSet::reorder(std::map<size_t, size_t> mapping) {
     for (auto &u: list) {
         in[u] = false;
         u = mapping[u];
@@ -141,11 +141,11 @@ void VertexSet::reorder(std::map<int, int> mapping) {
 void VertexSet::readFromText(const char *input_path) {
     FILE *input_file = fopen(input_path, "r");
 
-    int n, l;
-    fscanf(input_file, "%d%d", &n, &l);
-    std::vector<int> list(l);
-    for (int i = 0; i < l; ++i) {
-        int u; fscanf(input_file, "%d", &u);
+    size_t n, l;
+    fscanf(input_file, "%zu%zu", &n, &l);
+    std::vector<size_t> list(l);
+    for (size_t i = 0; i < l; ++i) {
+        size_t u; fscanf(input_file, "%zu", &u);
         list[i] = u;
     }
     fclose(input_file);
@@ -156,9 +156,9 @@ void VertexSet::readFromText(const char *input_path) {
 void VertexSet::writeToText(const char *output_path) const {
     FILE *output_file = fopen(output_path, "w");
 
-    fprintf(output_file, "%d %d\n", n, (int)size());
+    fprintf(output_file, "%zu %zu\n", n, size());
     for (auto u: list) {
-        fprintf(output_file, "%d ", u);
+        fprintf(output_file, "%zu ", u);
     }
     fputs("", output_file);
     fclose(output_file);
