+++ src/solvers/fista.h
@@ -9,8 +9,8 @@
 class Fista : public DSSolver{
 private:
 
-    long long m_R; // number of edges with an edge in R
-    int delta; // maximum vertex degree
+    size_t m_R; // number of edges with an edge in R
+    size_t delta; // maximum vertex degree
     double alpha; // learning rate
     std::vector<double> a;
     std::vector<double> r;
@@ -18,14 +18,14 @@ private:
     void initSol();
     void calcW();
     void calcR(const std::vector<double> &a);
-    int subgradient(std::vector<int> &r_argmx) const;
+    size_t subgradient(std::vector<size_t> &r_argmx) const;
     double learningRate(const std::vector<double> &nabla) const;
     double learningRateAsynchronous() const;
-    void gradient(const double &g, const std::vector<double> &a, std::vector<double> &nabla, const std::vector<int> &r_argmx, const int &r_nmx) const;
-    void descent(const std::vector<int> &r_argmx, const int &r_nmx, std::vector<double> &a_desc, int &t) const;
-    void descentAsynchronous(const std::vector<int> &r_argmx, const int &r_nmx, const std::vector<double> &x_old, std::vector<double> &x_new, std::vector<double> &y, std::vector<double> &r, const double &tk, const double &tk_new);
+    void gradient(const double &g, const std::vector<double> &a, std::vector<double> &nabla, const std::vector<size_t> &r_argmx, const size_t &r_nmx) const;
+    void descent(const std::vector<size_t> &r_argmx, const size_t &r_nmx, std::vector<double> &a_desc, int &t) const;
+    void descentAsynchronous(const std::vector<size_t> &r_argmx, const size_t &r_nmx, const std::vector<double> &x_old, std::vector<double> &x_new, std::vector<double> &y, std::vector<double> &r, const double &tk, const double &tk_new);
     void project(std::vector<double> &a_proj) const;
-    void fractionalPeeling(std::vector<int> &ord) const;
+    void fractionalPeeling(std::vector<size_t> &ord) const;
     void extractGraph(const double &rho);
 
 public:
